Refactor: SingMode AiAnswering.tsx 로직 분리
